add softdelete and fix database

+++ database/migrations/2025_04_21_154720_create_transactions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->unsignedBigInteger('user_id');
-            $table->decimal('total_price', 10, 2);
+            $table->integer('total_price');
             $table->timestamp('transaction_date')->useCurrent();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
